+++ CHANGELOG.md
@@ -1,4 +1,9 @@
-## v0.1 - April 4th, 2026
+## v0.2.0-prerelease - April 6th, 2026
+- Added pause/resume feature
+- Basic session information viewer
+- Various bug fixes
+
+## v0.1.0-prerelease - April 4th, 2026
 - Bracket validation for item tracking
 - Dehkia II zone logic
 - Zone logic detection for Mansha Forest / Cyclops and Sausan Garrison / Shultz Guard